refactor(codec): moving legacyAmino methods to codec package
