+++ go/properties/base/codec.go
@@ -3,7 +3,7 @@ package base
 import (
 	"github.com/cosmos/cosmos-sdk/codec"
 
-	codecUtilities "github.com/AssetMantle/schema/go/utilities"
+	codecUtilities "github.com/AssetMantle/schema/go/codec"
 )
 
 func RegisterLegacyAminoCodec(legacyAmino *codec.LegacyAmino) {
