+++ go/qualified/mutables.go
@@ -1,11 +1,6 @@
 // Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
 // SPDX-License-Identifier: Apache-2.0
 
-/*
- *  Copyright [2021] - [2022], AssetMantle Pte. Ltd. and the code contributors
- *  SPDX-License-Identifier: Apache-2.0
- */
-
 package qualified
 
 import (
